Move models folder to outside the app
