+++ src/app/review-essay/review-essay.component.ts
@@ -3,7 +3,7 @@ import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 
 import { ReviewService } from '../services/review.service';
-import { Review } from '../models/review';
+import { Review } from '../../models/review';
 
 import {BehaviorSubject} from 'rxjs/BehaviorSubject';
 import { Subscription } from 'rxjs/Subscription';
