+++ src/app/services/authentication.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 
-import {LoginBody, RegistrationBody, UpdatePassBody} from '../models/body-obj.model';
+import {LoginBody, RegistrationBody, UpdatePassBody} from '../../models/body-obj.model';
 import { environment } from '../../environments/environment';
 
 import 'rxjs/add/operator/map';
