+++ src/app/services/user.service.ts
@@ -3,7 +3,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 
 import { environment } from '../../environments/environment';
-import {EditUserBody} from '../models/body-obj.model';
+import {EditUserBody} from '../../models/body-obj.model';
 
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
